Sample intervals greater than 1000 record data, data point count division error fixed.

+++ iOS/Data_Collector/Classes/AutomaticViewController.mm
@@ -508,7 +508,7 @@ sampleInterval, geoCoder, city, address, country, dataSaver, managedObjectContex
         else
             secondsStr = [NSString stringWithFormat:@"%d", seconds];
         
-        int dataPoints = (1000 / recordingRate) * elapsedTime;
+        int dataPoints = (int) (1000 / ((float)recordingRate) * elapsedTime);
         
         NSLog(@"points: %d", dataPoints);
